rename model index and repository index
